+++ app/src/Factory/ProjectDtoFactoryInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Factory;
+
+use App\DataTransferObject\ProjectUpdateDto;
+use App\Entity\Project;
+
+interface ProjectDtoFactoryInterface
+{
+    public function createUpdate(Project $project): ProjectUpdateDto;
+}