+++ app/src/Form/ProjectCreateType.php
@@ -4,13 +4,12 @@ declare(strict_types=1);
 
 namespace App\Form;
 
-use App\DataTransferObject\BugCreateDto;
-use App\Entity\Project;
-use Symfony\Bridge\Doctrine\Form\Type\EntityType;
+use App\DataTransferObject\ProjectCreateDto;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-final class BugCreateType extends BugBaseType
+final class ProjectCreateType extends ProjectBaseType
 {
     /**
      * @param array<mixed> $options
@@ -21,18 +20,14 @@ final class BugCreateType extends BugBaseType
         $this->addBaseFields($builder);
 
         $builder->add(
-            'project',
-            EntityType::class,
-            [
-                'label' => 'bug.form.project',
-                'class' => Project::class,
-                'placeholder' => 'bug.form.project_placeholder',
-            ],
+            'projectId',
+            TextType::class,
+            ['label' => 'project.form.project_id'],
         );
     }
 
     public function configureOptions(OptionsResolver $resolver): void
     {
-        $resolver->setDefault('data_class', BugCreateDto::class);
+        $resolver->setDefault('data_class', ProjectCreateDto::class);
     }
 }