+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Update to Symfony 5.1.
 - Additional PHPStan rules from Ergebnis.
+- Project management (list/create/edit).
+- Vim configuration for Vdebug.
 
 ## [0.9.0] - 2020-04-09